Split C library API into its own module.
C++ API is entirely based on the C one.
